initial etl in python
Initial version of an ETL program to transform JSON data from collectd
and load it into couchdb.  It's currently synchronous.  Will need to
benchmark performance with an asynchrnous version and implementing it in
C.  Expecting extremely high performance requirements when used to
monitor PostgreSQL database with relations nearing 1 million objects.

Initially proofed with python 2.4 thus uses the simplejson package.
Will need to modify that when proofed with python 2.6 or later, when
json is part of the core libraries.