stop threads on ctrl-c
Now when ctrl-c is used, the threads still stop the next time they try
to dequeue anything.  Also use a random number generator to make threads
sleep up to 5 seconds if there is nothing to do to prevent a stampeding
herds issue.  Hopefully 1 second granularity is sufficient when hundreds
or thousands of threads are used.