use timezones and utc
collectd gives us ctime without timezone information.  The resulting
data will be inserted into the database in UTC time using the etl's
timezone as the point of reference from UTC.  This dictates that the etl
needs to be in the same time zone as the systems under monitor by
collectd.